+++ VM/instructions/comparisonJumps.cpp
@@ -267,7 +267,7 @@ void jump_if_zero(){
     Variable top = stack.top();
     stack.pop();
     if((top.getType() == FLOAT && top.getFloatValue() == 0.0f) || (top.getType() == INT && top.getIntValue() == 0)){
-        if(bytecode.at(programCounter).getTypeAsString() == "INT" && bytecode.size() > bytecode.at(programCounter).getIntFromMemory()){
+        if(bytecode.at(programCounter).getTypeAsString() == "INT" && bytecode.size() > static_cast<size_t>(bytecode.at(programCounter).getIntFromMemory())){
             int target_address = bytecode.at(programCounter).getIntFromMemory();
             // PC gets incremented by PiELo::step() finishing
             programCounter = target_address - 1;
@@ -284,7 +284,7 @@ void jump_if_not_zero(){
     Variable top = stack.top();
     stack.pop();
     if((top.getType() == FLOAT && top.getFloatValue() != 0.0f) || (top.getType() == INT && top.getIntValue() != 0)){
-        if(bytecode.at(programCounter).getTypeAsString() == "INT" && bytecode.size() > bytecode.at(programCounter).getIntFromMemory()){
+        if(bytecode.at(programCounter).getTypeAsString() == "INT" && bytecode.size() > static_cast<size_t>(bytecode.at(programCounter).getIntFromMemory())){
             int target_address = bytecode.at(programCounter).getIntFromMemory();
             // PC gets incremented by PiELo::step() finishing
             programCounter = target_address - 1;
